refactor and split tests
